Reorganize compilation code folder structure
