+++ gui/src/app/Compilation/Context/compileStanProgram.ts
