+++ gui/src/app/Compilation/Control/CompilationServerDialog.tsx
@@ -1,4 +1,4 @@
-import { CompileContext } from "@SpCompileContext/CompileContext";
+import { CompileContext } from "@SpCompilation/CompileContext";
 import { Refresh } from "@mui/icons-material";
 import Divider from "@mui/material/Divider";
 import FormControl from "@mui/material/FormControl";
@@ -10,11 +10,11 @@ import RadioGroup from "@mui/material/RadioGroup";
 import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
 import { FunctionComponent, useCallback, useContext } from "react";
+import { serverTypeForUrl } from "./CompilationServerToolbar";
 import {
   localCompilationServerUrl,
   publicCompilationServerUrl,
-  serverTypeForUrl,
-} from "./CompilationServerConnectionControl";
+} from "./Constants";
 
 type ConfigureCompilationServerDialogProps = {
   isConnected: boolean;
