+++ gui/src/app/Compilation/Control/Constants.ts
@@ -1,0 +1,4 @@
+export const publicCompilationServerUrl =
+  "https://stan-wasm.flatironinstitute.org";
+export const localCompilationServerUrl = "http://localhost:8083";
+export type ServerType = "public" | "local" | "custom";
